fix: focus on slide when properties shown from sidebar
